+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main)  [1.4.0]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main)  [1.5.0]
+
+### Modified
+- fix bug where filters don't work in remote list_datasets
+- change order of kwargs in list_datasets: filename now the first kwarg
+- can now filter by list of filename strings functioning as a logical OR
+- dataset kwarg renamed to filename in list_revisions
+- fix ALF regular expression pattern: attribute, timescale and extension now parsed correctly
+- can now filter datasets by timescale
+- clearer error message auth errors raised
+- alyx client is_logged_in method now a dependent property
+
+## [1.4.0]
 
 ### Modified